backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.8.6
chromadb==0.4.22
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16
pypdf==3.17.4
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
redis==5.0.1
celery==5.3.6
